requirements.txt
# =============================================================================
# LLM Performance Testing Framework - Dependencies
# =============================================================================

# Core Framework Dependencies
# ---------------------------
streamlit>=1.46.0                    # Web UI framework
streamlit-autorefresh>=1.0.1        # Auto-refresh functionality for Streamlit
pyyaml>=6.0.2                       # YAML configuration file parsing
python-dotenv>=1.0.1                # Environment variable management

# Data Processing & Analysis
# --------------------------
pandas>=2.2.2                       # Data manipulation and analysis
numpy>=1.26.4                       # Numerical computing
altair>=5.0.0                       # Statistical visualization

# LLM & AI Evaluation
# -------------------
deepeval>=2.4.8                     # LLM evaluation framework
langchain-core>=0.1.0               # LangChain core functionality
openai>=1.65.2                      # OpenAI API client

# HTTP & API Communication
# ------------------------
requests>=2.32.3                    # HTTP library for API calls

# =============================================================================
# Development Dependencies (uncomment if needed)
# =============================================================================
# pytest>=7.0.0                    # Testing framework
# pytest-asyncio>=0.21.0           # Async testing support
# black>=23.0.0                     # Code formatter
# flake8>=6.0.0                     # Linting
# mypy>=1.0.0                       # Type checking
# pre-commit>=3.0.0                 # Git hooks for code quality
